Move files close to the source; Decouple some components
