+++ application/ui/src/components/icon-wrapper/icon-wrapper.component.tsx
