+++ application/ui/src/components/icon-wrapper/icon-wrapper.module.scss
