+++ application/ui/src/features/annotator/annotator-canvas.tsx
@@ -3,15 +3,19 @@
 
 import { View } from '@geti/ui';
 import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
+import { API_BASE_URL } from 'src/api/client';
 
 import { ZoomTransform } from '../../components/zoom/zoom-transform';
-import { getImageUrl } from '../dataset/gallery/utils';
 import { useAnnotationActions } from './annotation-actions-provider.component';
 import { useAnnotationVisibility } from './annotation-visibility-provider.component';
 import { Annotations } from './annotations/annotations.component';
 import { useSelectedAnnotations } from './select-annotation-provider.component';
 import { ToolManager } from './tools/tool-manager.component';
 import { DatasetItem } from './types';
+
+const getImageUrl = (projectId: string, itemId: string) => {
+    return `${API_BASE_URL}/api/projects/${projectId}/dataset/items/${itemId}/binary`;
+};
 
 type AnnotatorCanvasProps = {
     mediaItem: DatasetItem;
