+++ application/ui/src/features/annotator/tools/annotator-tools.component.tsx
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Divider } from '@geti/ui';
-import { BoundingBox, SegmentAnythingIcon, Selector } from '@geti/ui/icons';
+import { BoundingBox, Polygon, SegmentAnythingIcon, Selector } from '@geti/ui/icons';
 import { useProjectTask } from 'hooks/use-project-task.hook';
 
-import { ToolConfig } from '../../dataset/media-preview/primary-toolbar/tools/interface';
-import { Tools } from '../../dataset/media-preview/primary-toolbar/tools/tools.component';
 import { useAnnotator } from '../annotator-provider.component';
+import { ToolConfig } from './interface';
+import { Tools } from './tools.component';
 
 const TASK_TOOL_CONFIG: Record<string, ToolConfig[]> = {
     classification: [],
@@ -18,7 +18,7 @@ const TASK_TOOL_CONFIG: Record<string, ToolConfig[]> = {
     instance_segmentation: [
         { type: 'selection', icon: Selector },
         { type: 'sam', icon: SegmentAnythingIcon },
-        // TODO: Add 'polygon' tool later
+        { type: 'polygon', icon: Polygon },
     ],
 };
