+++ application/ui/src/features/annotator/tools/interface.ts
