+++ application/ui/src/features/dataset/media-preview/primary-toolbar/hotkeys/hotkeys.component.tsx
@@ -3,7 +3,7 @@
 
 import { Hotkeys as HotkeysIcon } from '@geti/ui/icons';
 
-import { IconWrapper } from '../icon-wrapper.component';
+import { IconWrapper } from '../../../../../components/icon-wrapper/icon-wrapper.component';
 
 export const Hotkeys = () => {
     return (
