+++ application/ui/src/features/dataset/media-preview/primary-toolbar/zoom/zoom-selector.component.tsx
@@ -4,8 +4,8 @@
 import { ActionButton, Flex, Text } from '@geti/ui';
 import { Add, Remove } from '@geti/ui/icons';
 
+import { IconWrapper } from '../../../../../components/icon-wrapper/icon-wrapper.component';
 import { useSetZoom, useZoom } from '../../../../../components/zoom/zoom.provider';
-import { IconWrapper } from '../icon-wrapper.component';
 
 export const ZoomSelector = () => {
     const zoom = useZoom();
